refactor: init sandbox
and mv ResourcesEnum to resources
